AG-1073: bug fixes - override existing options when showing no data chart, ensure x-axis tooltip labels update when mapping is updated
