+++ libs/shared/typescript/charts/src/lib/x-axis-label-tooltips/index.ts
@@ -1,0 +1,1 @@
+export * from './x-axis-label-tooltips';
